Update version 0.1.1-SNAPSHOT.
commit dde54b61f6641d933d2f9073d8aa617d8d075a55
Author: Toshiki Takeuchi <take@xcoo.jp>
Date:   Fri Mar 30 21:44:03 2012 +0900

    Add CMYKColor class.
    Add 2 enum classes for texture.

commit 5afbef4651220c3790fc5705da74aa1af78d3319
Author: Toshiki Takeuchi <take@xcoo.jp>
Date:   Fri Mar 30 01:41:39 2012 +0900

    Integrate mouseWheelEvent with mouseEvent.

commit 4aab2d45d5f9328927fab01240b89f572f9081ec
Author: Toshiki Takeuchi <take@xcoo.jp>
Date:   Thu Mar 22 20:24:17 2012 +0900

    Fix a bug of SQL.
    Modify ArrayList of GraphicsObject to CopyOnWriteArrayList.

commit 080c5192978a4cbf710447a73c9e80d87fb510d6
Author: Toshiki Takeuchi <take@xcoo.jp>
Date:   Tue Mar 20 12:19:39 2012 +0900

    Update pom.xml to use latest OpenNI and NITE.

commit 9ace0525d9f4d14a21704df1a71dc721c9cc0487
Author: Toshiki Takeuchi <take@xcoo.jp>
Date:   Tue Mar 20 10:41:01 2012 +0900

    Modify Trackball.

commit 78ab5ddfdd84e748131de12487e4fe6cbc4a0f5f
Author: Toshiki Takeuchi <take@xcoo.jp>
Date:   Fri Mar 16 00:49:27 2012 +0900

    Add Mouse class.

commit a3152207d195bf41af6369b2beadd4f99d16442a
Author: Toshiki Takeuchi <take@xcoo.jp>
Date:   Thu Mar 15 16:13:46 2012 +0900

    Fix a reshape bug.

commit 396afe9e46e2e505b45303a40de89d32db12f12e
Author: Toshiki Takeuchi <take@xcoo.jp>
Date:   Sat Mar 10 18:32:31 2012 +0900

    Add reset method to Trackball class.
    Refactoring Applet and CursorMode classes.

commit 1bfbab9ddc3c23ffa4cfaf653a179dda062537e6
Author: Toshiki Takeuchi <take@xcoo.jp>
Date:   Sat Mar 10 15:16:34 2012 +0900

    Add Trackball class.

commit 7e113c4d3fadf7df01dcadd472f471a00db5c9d5
Author: Toshiki Takeuchi <take@xcoo.jp>
Date:   Fri Mar 9 14:37:34 2012 +0900

    Fix Box class.

commit 01fe299dbb2ef0c2ef24b1da46c3888d559edaa1
Merge: 5022846 55ed6df
Author: Yuki Ban <yuki@xcoo.jp>
Date:   Wed Mar 7 09:09:37 2012 +0900

    Merge branch 'dev' of github.com:casmi/casmi into dev

commit 50228467f0fb686e92ce517772c4e5372cff5cef
Author: Yuki Ban <yuki@xcoo.jp>
Date:   Wed Mar 7 09:09:28 2012 +0900

    Modified TimeLine class

commit 55ed6df1f21a258ec3ced88677de2d08ca5cb2cf
Author: Toshiki Takeuchi <take@xcoo.jp>
Date:   Tue Mar 6 14:22:05 2012 +0900

    Remove warnings of CPopup.

commit d5663c385afd986f24ce55fed7ac64b02193452b
Author: Toshiki Takeuchi <take@xcoo.jp>
Date:   Fri Mar 2 14:51:54 2012 +0900

    Add license and javadoc comments.

commit 096fec8a5adf5504068eac0b4fba49275c09009b
Author: Yuki Ban <yuki@xcoo.jp>
Date:   Fri Mar 2 14:35:54 2012 +0900

    Modified selection buffer

commit 3cf0da85cd3259f38309dc1f5463f0f0abdc74e3
Author: XcooDev <developer@xcoo.jp>
Date:   Sat Feb 25 23:37:00 2012 +0900

    add synchronized block in rendering

commit b05135396611aa4409d9ca1690f6331604427aad
Merge: 4df978d 8ba4300
Author: XcooDev <developer@xcoo.jp>
Date:   Sat Feb 25 22:28:24 2012 +0900

    Merge branch 'dev' of github-xcoo:casmi/casmi into dev

commit 4df978d2d072a24bced11e61b192a8515c2de92c
Author: XcooDev <developer@xcoo.jp>
Date:   Sat Feb 25 22:28:02 2012 +0900

    add function to show a alert dialog box

commit 8ba430072ab801e1b1ed1ce337cb1d5b8424e4eb
Author: Toshiki Takeuchi <take@xcoo.jp>
Date:   Sat Feb 25 21:24:01 2012 +0900

    Refactoring some classes.

commit 2e250a563c4efc9e2b2abb892ebb266ef6e7a93b
Author: XcooDev <developer@xcoo.jp>
Date:   Sat Feb 25 19:26:10 2012 +0900

    fix bug in Polygon class

commit c29f1f1167281c5fd306618b2c76a25675872275
Author: Toshiki Takeuchi <take@xcoo.jp>
Date:   Sat Feb 25 17:54:35 2012 +0900

    Add GrayColor class.

commit 1f7a40470439fda330cffb15bce51bbd282dd92a
Author: XcooDev <developer@xcoo.jp>
Date:   Sat Feb 25 17:53:28 2012 +0900

    fix bug in Line class

commit 578f12c89869ab394a5595817e3fd0197890d21c
Author: Yuki Ban <yuki@xcoo.jp>
Date:   Wed Feb 22 15:23:57 2012 +0900

    add Mouse Wheel Event.

commit e46b2a55bc08505f11ff6f64533c861f61defef7
Author: shigeodayo <shigeo.yoshida.shigeo@gmail.com>
Date:   Fri Feb 17 14:59:17 2012 +0900

    Add CPopupmenu

    Add CPopupmenu, easy to create popup menu.

commit ce5bf19848573173440f03a0dc08da8cfd1e8701
Author: Toshiki Takeuchi <take@xcoo.jp>
Date:   Fri Feb 17 13:18:33 2012 +0900

    Fix bugs about Primary Key and Short/Long type.

commit 18c992fe2121a6538f0aa33a8035ff0af1f516c3
Author: Toshiki Takeuchi <take@xcoo.jp>
Date:   Tue Feb 14 05:11:16 2012 +0900

    Divide Color class to three classes (Color interface, RGBColor class, HSBColor class).
    Remove ColorMode.
    Refactoring many sources.

commit 469f0e6cd6d5985e799e54ca87892c2620685872
Author: XcooDev <developer@xcoo.jp>
Date:   Thu Feb 9 17:29:17 2012 +0900

    fixed version reference in README

commit 814de6bf236f866956003c9e72073ca5caf26103
Author: Toshiki Takeuchi <take@xcoo.jp>
Date:   Sat Feb 4 17:07:36 2012 +0900

    Update version to 0.1.1-SNAPSHOT.

commit 58b4744ef60cc435965aec28e40c639948637270
Author: Toshiki Takeuchi <take@xcoo.jp>
Date:   Sat Feb 4 16:46:46 2012 +0900

    Rename Roundrect.java to RoundRect.java.
    Add maven-compiler-plugin version to pom.xml.

commit b31e6ae4be80737254690e05086121de8652a955
Author: Yuki Ban <yuki@xcoo.jp>
Date:   Sat Feb 4 15:16:20 2012 +0900

    Add repeat function to TweenGroup

commit 50c6950486df6bbf13a6c6c236ce2fffe2b5f85e
Author: Toshiki Takeuchi <toshiki@Prima-Alice.local>
Date:   Sat Feb 4 14:31:09 2012 +0900

    Add a repository icon.

+++ src/main/java/casmi/sql/StatementGenerator.java
@@ -66,10 +66,10 @@ class StatementGenerator {
         TYPE_TABLE.put(key1(SQLType.MYSQL_5,           null),     "NULL");
         TYPE_TABLE.put(key1(SQLType.MYSQL_5,      int.class),      "INT");
         TYPE_TABLE.put(key1(SQLType.MYSQL_5,  Integer.class),      "INT");
-        TYPE_TABLE.put(key1(SQLType.MYSQL_5,    short.class),      "INT");
-        TYPE_TABLE.put(key1(SQLType.MYSQL_5,    Short.class),      "INT");
-        TYPE_TABLE.put(key1(SQLType.MYSQL_5,     long.class),      "INT");
-        TYPE_TABLE.put(key1(SQLType.MYSQL_5,     Long.class),      "INT");
+        TYPE_TABLE.put(key1(SQLType.MYSQL_5,    short.class), "SMALLINT");
+        TYPE_TABLE.put(key1(SQLType.MYSQL_5,    Short.class), "SMALLINT");
+        TYPE_TABLE.put(key1(SQLType.MYSQL_5,     long.class),   "BIGINT");
+        TYPE_TABLE.put(key1(SQLType.MYSQL_5,     Long.class),   "BIGINT");
         TYPE_TABLE.put(key1(SQLType.MYSQL_5,    float.class),    "FLOAT");
         TYPE_TABLE.put(key1(SQLType.MYSQL_5,    Float.class),    "FLOAT");
         TYPE_TABLE.put(key1(SQLType.MYSQL_5,   double.class),   "DOUBLE");
